requirements.txt
# Web Framework
Flask==3.0.3
Flask-Login==0.6.3

# Excel/CSV export
openpyxl==3.1.5

# Environment variables
python-dotenv==1.0.0

# AI/ML
openai==1.12.0

# AWS Services
boto3==1.34.0

# Security (for password hashing - included with Werkzeug/Flask)
